добавил синглтоны, переделал метод object, чуть подправил код
