+++ ImageFeed/Network/UIBlockingProgressHUD.swift
@@ -2,8 +2,6 @@
 //  UIBlockingProgressHUD.swift
 //  ImageFeed
 //
-//  Created by Гость on 12.06.2023.
-//
 
 import UIKit
 import ProgressHUD
